pyproject.toml
[project]
name = "swarm-ai"
version = "1.0.0"
description = "A hive-mind for Claude Code agents — TUI + web orchestrator"
readme = "README.md"
authors = [
    { name = "Brad Schleifer", email = "bschleifer@rcg.org" }
]
requires-python = ">=3.12"
dependencies = [
    "textual>=1.0",
    "textual-serve>=1.1",
    "click>=8.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "aiohttp-jinja2>=1.6",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
swarm = "swarm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/swarm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "C901"]

[tool.ruff.lint]
select = ["E", "F", "W", "C901"]
